// frontend/src/app/api/model/languageDto.ts
/**
 * Book Shelf API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface LanguageDto {
  /**
   * Unique language identifier.
   */
    id?: string;
  /**
   * Language name.
   */
    name?: string;
}