// frontend/src/app/api/model/pagedResponseLanguageDto.ts
/**
 * OpenAPI definition
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import {LanguageDto} from './languageDto';


export interface PagedResponseLanguageDto {
  content?: Array<LanguageDto>;
  pageNumber?: number;
  pageSize?: number;
  totalElements?: number;
  totalPages?: number;
  isLast?: boolean;
}